rename AktiviteterTilKafka pakken (#158)
renamer KafkaProducerService sin record til kafkaMelding java:S6213
