+++ src/main/java/no/nav/veilarbaktivitet/aktiviteter_til_kafka/EndringsType.java
@@ -1,4 +1,4 @@
-package no.nav.veilarbaktivitet.aktiviteterTilKafka;
+package no.nav.veilarbaktivitet.aktiviteter_til_kafka;
 
 import no.nav.veilarbaktivitet.domain.AktivitetTransaksjonsType;
