+++ src/main/java/no/nav/veilarbaktivitet/service/CronService.java
@@ -3,7 +3,7 @@ package no.nav.veilarbaktivitet.service;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import no.nav.common.job.leader_election.LeaderElectionClient;
-import no.nav.veilarbaktivitet.aktiviteterTilKafka.AktiviteterTilKafkaService;
+import no.nav.veilarbaktivitet.aktiviteter_til_kafka.AktiviteterTilKafkaService;
 import no.nav.veilarbaktivitet.motesms.MoteSMSService;
 import org.slf4j.MDC;
 import org.springframework.scheduling.annotation.EnableScheduling;
