+++ src/test/java/no/nav/veilarbaktivitet/db/dao/KafkaAktivitetDAOTest.java
@@ -1,7 +1,7 @@
 package no.nav.veilarbaktivitet.db.dao;
 
 import lombok.val;
-import no.nav.veilarbaktivitet.aktiviteterTilKafka.KafkaAktivitetMeldingV4;
+import no.nav.veilarbaktivitet.aktiviteter_til_kafka.KafkaAktivitetMeldingV4;
 import no.nav.veilarbaktivitet.db.Database;
 import no.nav.veilarbaktivitet.db.DbTestUtils;
 import no.nav.veilarbaktivitet.domain.AktivitetData;
