+++ src/test/java/no/nav/veilarbaktivitet/domain/EndringsTypeTest.java
@@ -1,6 +1,6 @@
 package no.nav.veilarbaktivitet.domain;
 
-import no.nav.veilarbaktivitet.aktiviteterTilKafka.EndringsType;
+import no.nav.veilarbaktivitet.aktiviteter_til_kafka.EndringsType;
 import org.junit.Test;
 
 import static org.assertj.core.api.Assertions.assertThat;
