+++ src/test/java/no/nav/veilarbaktivitet/service/KafkaProducerServiceTest.java
@@ -1,7 +1,7 @@
 package no.nav.veilarbaktivitet.service;
 
 import no.nav.common.kafka.producer.KafkaProducerClient;
-import no.nav.veilarbaktivitet.aktiviteterTilKafka.KafkaAktivitetMeldingV4;
+import no.nav.veilarbaktivitet.aktiviteter_til_kafka.KafkaAktivitetMeldingV4;
 import no.nav.veilarbaktivitet.config.KafkaProperties;
 import org.apache.kafka.clients.producer.ProducerRecord;
 import org.apache.kafka.clients.producer.RecordMetadata;
